Conflicts:
	corlib/System.Collections.IList.h
	corlibtest/corlibtest.vcxproj.filters

+++ corlib/System.Array.hpp
@@ -3,6 +3,7 @@
 #pragma once
 
 #include "System.Array.h"
+#include "System.Array.hpp"
 #include "System.Exception.h"
 #include "System.NotImplementedException.h"
 #include "System.Collections.Generic.EqualityComparer.h"
@@ -215,7 +216,7 @@ namespace System
     if(startIndex < arr.GetLowerBound() || startIndex - 1 > arr.GetUpperBound() - count)
       throw ArgumentOutOfRangeException ();
 
-    return Generic::EqualityComparer<T>::Default()->IndexOf(arr, value, startIndex, startIndex + count);
+    return (int)Generic::EqualityComparer<T>::Default()->IndexOf(arr, value, (int)startIndex, (int)(startIndex + count) );
     }
 
   }
